Added install/uninstall batch files, test local admin in uninstall

+++ Invoke-EnvironmentTeardown.ps1
@@ -6,6 +6,19 @@ $Global:win10sdkPath = "$Global:psenvPath\win10sdk"
 $Global:opensshPath = "$Global:psenvPath\openssh"
 $Global:setupScriptsPath = "$Global:psenvPath\scripts\setuphelpers"
 
+function Test-Administrator {
+    $identity = [System.Security.Principal.WindowsIdentity]::GetCurrent()
+    $principal = New-Object System.Security.Principal.WindowsPrincipal($identity)
+    $admin = [System.Security.Principal.WindowsBuiltInRole]::Administrator
+    $IsAdmin = $principal.IsInRole($admin)
+    return $IsAdmin
+}
+
+if (!(Test-Administrator)) {
+    Write-Warning -Message "You are not running as a local admin; please start this as a local admin!"
+    exit
+}
+
 # Delete Profile
 Write-Host "Deleting profile..." -ForegroundColor Cyan
 Remove-Item -Path $profile.AllUsersAllHosts
